api-fort-43 Etl to Dev and Dev to Prod processes should delete rows from dev (or prod)
#what

Sets the records in the dev-performer table to have a wave of 9999 if the record found in the dev table is not in the etl-performer database.

#why

The etl process was transferring data from one table to the next creating duplicates and allowing some performers that were deleted to be propagated from one table to the next. Instead of truncating the table (because there's no simple way to truncate, you have to delete, then create then wait for the table to finish becoming available before inserting) the value for the wave is set to 9999. In our performer filter code we make sure to filter out any performers with a wave that is greater than the current_wave_number. This in turn acts as a soft delete for right now.

This change is not applied to prod as prod should directly reflect dev once prod has been loaded. I also ran this script on prod so now all artists should reflect what is in etl.